requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==41.0.4
Django==4.2.5
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.5
phonenumbers==8.13.21
Pillow==10.0.1
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==7.6.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
